ppf-route-api/requirements.txt
asgiref==3.7.2
boto3==1.34.108
botocore==1.34.108
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-filter==24.2
django-storages==1.14
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==24.2.0
google-api-core==2.17.1
google-api-python-client==2.129.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-geo-type==0.3.7
google-maps-routing==0.6.7
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
haversine==2.8.1
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
geopy==2.2.0
polyline==1.4.0
scikit-learn==1.4.2
numpy==1.26.4
firebase-admin==6.5.0
django-storages==1.14.0
polyline==1.4.0
boto3
pillow

wheel==0.43.0